Move all services to com.twilio.chat.demo.services
