+++ chat-demo-android/src/main/AndroidManifest.xml
@@ -51,7 +51,7 @@
         <!-- FCM service -->
 
         <service
-            android:name=".FCMListenerService">
+            android:name=".services.FCMListenerService">
             <!-- android:exported="true" -->
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT"/>
@@ -68,7 +68,7 @@
         <!-- FCM instance ID service -->
 
         <service
-            android:name=".FCMInstanceIDService"
+            android:name=".services.FCMInstanceIDService"
             android:exported="true">
             <intent-filter>
                 <action android:name="com.google.firebase.INSTANCE_ID_EVENT"/>
@@ -76,7 +76,7 @@
         </service>
 
         <service
-            android:name=".RegistrationIntentService"
+            android:name=".services.RegistrationIntentService"
             android:exported="false">
         </service>
