+++ chat-demo-android/src/main/java/com/twilio/chat/demo/services/FCMInstanceIDService.kt
@@ -1,4 +1,4 @@
-package com.twilio.chat.demo
+package com.twilio.chat.demo.services
 
 import android.content.Intent
